[GUI] Add text size & line height calculation
These APIs are quite ugly, but they're a nice start. The problem is that
we'd like to have multithreading in order to have a smooth user
experience.

To achieve this, we need to have Send'able Painter / TextCalculator, but
the mltg library in combination with windows-rs and DirectX doesn't
allow this currently. Because of this, we need quite a few hacks,
including having a separate Direct2D context on the background thread.