[App] Display tab loading percentage
This way, the user has some indicating of how much the tab is loaded,
and how much more time it will take.

A time indicating (e.g. 5 seconds remaining) would be very aggressive,
and something we shouldn't want. If the loading is so extremely slow
that we need this, we have other problems at hand. ;)