*WIP* Start working on using thunderdome as the allocation strategy
Using thunderdome's arena/slab-esque allocator, we can allocate the
Nodes more performantly, since they all share the same size.

Since thunderdome is an arena-like allocator, we can also use the
indices of the nodes as references to the parent and child nodes. This
does however come at the cost of losing the non-nullness and dangling
reference prevention which the Vec<Node>, Vec<&Node> and
Vec<Rc<RefCell<Node>>> provided.

But this way, the allocation is faster, and (de)referencing the indices
is faster than using reference counting.